add authentication to goal routes

+++ controlles/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler =require('express-async-handler')
 const Goal =require('../models/goalModel');
+const User =require('../models/userModel')
 
 // @desc    Get Goals
 // @routes  GET /api/goals
 // @access  Private
 
 const getGoals =asyncHandler( async (req, res) => {
-  const goals =await Goal.find()
+  const goals =await Goal.find({user:req.user.id})
   res.status(200).json(goals);
 });
 
@@ -22,7 +23,8 @@ const setGoals =asyncHandler( async (req, res) => {
     throw new Error("Please add a Text");
   }
   const goal =await Goal.create({
-    text:req.body.text
+    text:req.body.text,
+    user:req.user.id
   })
 
   res.status(200).json(goal);
@@ -39,6 +41,23 @@ const updateGoals =asyncHandler( async (req, res) => {
     res.status(400)
     throw new Error('Goal not found')
   }
+
+  const user =await User.findById(req.user.id)
+
+  //check user is in the database
+  if(!user){
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  //make sure user delete his own goal
+
+  if(goal.user.toString() !==user.id){
+    res.status(401)
+    throw new Error("User not Authorized")
+  }
+
+
   const updatedgoal =await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
 
   res.status(200).json(updatedgoal);
@@ -54,6 +73,23 @@ const deleteGoals =asyncHandler( async (req, res) => {
     res.status(400)
     throw new Error('Goal not found')
   }
+
+  const user =await User.findById(req.user.id)
+
+  //check user is in the database
+  if(!user){
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  //make sure user delete his own goal
+
+  if(goal.user.toString() !==user.id){
+    res.status(401)
+    throw new Error("User not Authorized")
+  }
+
+  
  await Goal.findByIdAndRemove(goal._id)
   res.status(200).json({ id:  req.params.id });
 });
